The player can now properly handle headset control inputs.

+++ src/jni/src/SUI/ListView.cpp
@@ -119,7 +119,7 @@ update_restart:
 		SDL_SetRenderDrawColor(renderer.get(), 0xFF, 0xFF, 0xFF, 0x80);
 		SDL_Rect rect = {
 			this->visible_region.w - 10,
-			-this->offset / this->total_length * this->visible_region.h,
+			(int)(-this->offset / this->total_length * this->visible_region.h),
 			10,
 			this->visible_region.h * this->visible_region.h / this->total_length,
 		};
